requirements.txt
python-chess
transformers
torch
stable-baselines3[extra]
gym
pygame
svglib
reportlab